requirements.txt
# Core AI framework
pydantic-ai>=0.0.9
openai>=1.0.0

# Web interface
streamlit>=1.28.0

# Vector database
chromadb>=0.4.18

# Environment management
python-dotenv>=1.0.0

# Embeddings
sentence-transformers>=2.2.2

# Document processing
pypdf>=3.17.0
PyPDF2>=3.0.1
langchain-community>=0.0.1

# HTTP client for Ollama
httpx>=0.25.0
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Text processing
nltk>=3.8.0

# Utilities
pathlib>=1.0.1
typing-extensions>=4.8.0

# Async support
asyncio>=3.4.3
aiohttp>=3.9.0